Removed then blocks and it now works?
The problem with the while-then thing was that to make the then block
useful (at least without effects) would be to return a value, but that
is only safely possible if there is also a default value returned if
the loop does not run at all. This is fixable by also adding an else
clause, but then the syntax becomes a bit too much:
`while cond() {...} then {...} else {...}`
is a lot.

A solution could be to make the then clause not be it's own thing and
rather integrating it into the main body, returning whatever the
loop's body returns at the last iteration, given a default value is
provided by an else clause.
The problem then becomes that you cannot have it execute a different
body after the last iteration like so:
```tara
while cond() {
    print("printed in the loop");
} then {
    print("printed after the loop");
} else {
    print("printed if the loop never ran");
}
```
becomes impossible conceptually.